chore: 주석 변경 및 파일 분리
