+++ BoxOffice/Protocol/NetworkingProtocol.swift
@@ -2,12 +2,13 @@
 //  NetworkingProtocol.swift
 //  BoxOffice
 //
-//  Created by Andrew on 2023/03/28.
+//  Created by Andrew, 레옹아범 on 2023/03/28.
 //
 
 import Foundation
 
 protocol NetworkingProtocol {
-    func search(url: URL, completion: @escaping (Result<Data, BoxofficeError>) -> Void) -> URLSessionDataTask
     init(session: URLSession)
+    
+    func search(url: URL, completion: @escaping (Result<Data, BoxofficeError>) -> Void) -> URLSessionDataTask
 }
