+++ BoxOffice/SceneDelegate.swift
@@ -2,7 +2,7 @@
 //  SceneDelegate.swift
 //  BoxOffice
 //
-//  Created by kjs on 13/01/23.
+//  Created by Andrew, 레옹아범 on 13/01/23.
 //
 
 import UIKit
